Working evil-paredit key mapping
The previous one was rebinding everywhere, which was causing problems
when editing org-mode headings... yay conflicts!

Need to use evil's keybinding function in order to bind within this
specific mapping AND specific evil-mode (e.g. this definition of
backspace only should be mapped when evil-paredit is on, and should
only apply to insert-mode)