Switching file-find back to ido
I never quite got on with helm for finding files... found ido to be
much much more in tune with "how I think" which I guess is the most
important thing of all.

That said, I wouldn't be averse to dragging along helm for things
like helm-dash... except that I've traditionally not used them and
have preferred just looking it up in my web browser.

I'll have to see.