Removed the old helm.el, using some of helm
As mentioned in the source comment, I like helm's ancillary functions
like 'helm-colors, but I can't stand 'helm-find-files, it just doesn't
mesh with how I think about finding files.

The old init-helm.el was an attempt to hack ido-like file finding into
helm, but it was just that... a hack.  Might as well just use both with
a preference for ido.