magit just pushed a huge change; trying to fix
This must have been a full point release (like this 1.4 -> 2.0 that I'm
seeing) because I don't normally see this amount of breakage.

You need to have git >=1.9.4.  I'm usually on Ubuntu 14.04 LTS, which is
still stuck on 1.9.1 by default.  The following instructions got this
Ubuntu box up to... wow, they're at 2.4.5 now??  Anywho, this worked for
me:

https://www.howtoforge.com/install-git-on-ubuntu-13.1

Upgrading git and upgrading these config files to this commit should,
along with manually removing git-commit-mode and git-revert-mode
(IIRC, it should barf out which modes need to be removed in the error on
startup x_x), actually get emacs to start up properly.

I am now noticing problems with the evil mappings in various magit
modes.  It's workable (so long as you use C-p/C-n instead of jk) so just
keep mindful of that.