Finally made nth-selection in company-mode work
For whatever reason, M-1, M-2, etc are broken on my setup because Meta
and Escape map to very similar keys or something underneath... sometimes
it's a bit hard to keep the various legacy terminal keyboard settings
straight.  In any case, I just mapped to C-1, C-2 when the company popup
is active.