Fix annoying desktop-mode code for abnormal exit
If emacs is killed abnormally (which does happen to me; I tend to run
bleeding-edge releases), I don't want to have to come back a day or
more later and find that I need to answer desktop-save-mode's annoying
bootup message.  I want to start coding!  So I put this hack in, which
I found from emacswiki as indicated in the file.