ivy is an almost perfect merging of ido and helm
with a heck of a lot less hairy configuration than either.

A lot of the below is very subjective; I think helm is a neat
package too, and ivy owes a lot to it.

ido:
  + comes with emacs
  + fast, both in performance and in keystrokes
  + with flx, pretty nice completion
  - needs a ton of other people's packages to do much
  - interface tries to not be super intrusive, which means it is
    limited in what it can do sometimes... without more package

helm:
  + does everything and the kitchen sink
  + tries to be a unified interface to finding things
  + did i mention that it can do everything?
  - does not come with emacs
  - feels heavy
  - even after significant refactoring and TLC, it still needs a
    lot of work and research to configure
  - keybindings not the most intuitive
  - completion method a little alien to those who like to
    tab-complete

ivy:
  + sort of like a "unifying interface" like helm
  + but works more like ido
  + feels faster than helm
  + much less configuration needed than either
  + feels "more modern" than helm
  - does not come with emacs
  - some keybindings still a little wonky (when I press RET on a
    directory, I don't want dired, I want to go into the directory)

I get that helm/ivy want to be this unifying interface, but I really
wish they'd break their rules with *-find-files and make RET mean
"open the file if it is a file, otherwise open helm/ivy on this
directory" and not "open the file, and if it is a directory, open it
in dired"