Hacky usage of spaceline-all-the-icons
I like spaceline-all-the-icons but you need to install icon fonts for
it to look right, and at the moment I'm not aware of a way to check
for the existence of these fonts.

And you lose nyan-mode T_T  If I ever get a wild hair, I might see what
needs to be done to get nyan-mode to work.

Right now, I just have the old smart-mode-line as default, and then
the user can require 'jws-modeline-fancy if they want.