If I wanted my hands held, I would use MS Word
Actually, if I wanted my hands held, I'd be working in sales rather
than programming... *sad trombone*

<rant>

Emacs is hard.  It is a decades-old text editor.  I get trying to be
friendly to new users, I really do.  It is death to a project to not
be open to newbies.  But I fail to see why opening a folder in the
same dired buffer is "confusing."  The old behavior was confusing and
infuriating to me, because it makes ten zillion buffers as I navigated
my FS, when really all I wanted was just the one.

Argh!!!!

</rant>