"Clean up" org-mode config a bit
On my home computers, I run flycheck on elisp files.  This is good for
developing packages but less good for my actual emacs config, which I
tend to leave messy :P  One of the main reasons I haven't pushed these
elisp flycheck settings to github, actually, I want to be able to hack
this config and not have to worry about red squiggly lines.  It's not
like anyone else is using this, haha.