Revamped formatting options
The formatting options were moved into a hydra, which I'm going to
need to play around with a bit more.  I'm not sure I like :exit nil
(at least for this application) and need to work with it a bit more.
But it's a lot better than my old function that tried to handle this.

Other important things were moved here.  electric-indent-mode is
invaluable on modern code that is pretty closely formatted to
standards; but it can be infuriating otherwise, so it is an integral
part of the formatting hydra, as well as whitespace-mode.