Added "The BufJuggler"
There is a Vimscript called 'LustyJuggler' (I don't like the name
either...) that would map the 10 MRU buffers to the home row and the
number keys.  So, to switch to the secondmost recently used buffer,
you'd invoke LJ and then press 2 or s; to the tenth-most, you'd press
0 or ;.

This is an ersatz Emacs reimplementation of the concept, using hydra
as its GUI.  Basically, I cribbed the code at
https://github.com/abo-abo/hydra/wiki/Switch-to-buffer (which was like
85% of the way there already) and then hacked it till I had something
that semi-worked.

My elisp-fu is not great and hydra seems to have problems with some
aspect of my setup (hints don't always show up like they should if I
use a sexp, which is why I did the awful docstring thing which does
not align well), but I'd like to revisit the concept and see if I can
make it more customizable and nicer looking and then extract that as a
plugin.