Made window positioning "stupid"
The idea is to play nicer with emacs' daemon mode, which is otherwise
an unmitigated "Win" in terms of productivity, but messes with
anything that needs a GUI frame during initialization,
including... querying fonts, for some reason.  Which is what my
"smart" setup relied on.  Shame.

As a result, I've moved the "smarts" into an interactive command
intended to be run by the user from a GUI frame.  Once that is done,
the user can manually enter their preferred settings in their
site-init.el.