Juggle avy/company bindings; add yasnippet
I am going to try to get used to not using TAB to complete.  TAB is
already set in company as company-template-forward-field and I like it
there too...

If there was a company-template-backwards-field that would be great;
I'd set backtab to that.

To that end, C-i now completes common substring and then cycles, which
is what TAB used to do; C-j now completes with what's there.

Outside of the company-mode popup:

C-; now does a yasnippet expansion, or shows possible ones in a list
via company-yasnippet

C-' now jumps to character using avy was C-;)

C- now jumps to line using avy (was C-:)