Slowly starting to deprecate unneeded hydras
Hydras are pretty nice if you need the exact kind of interaction they
provide -- more-or-less static menu display and a way to provide a
guide for repeated button presses.  But if your needs involve:

 - adding more keys to your hydra dynamically (for instance, a menu
   of related things that might be defined in separate files)

 - having the elements of the hydra change dynamically (e.g. MRU file
   pickers like the BufJuggler I've been working on)

it isn't the easiest library to work with.  Either that or my brain
just doesn't work that way.

For my application menus, I've discovered how easy
define-prefix-command is to work with; it creates a map, and then you
can add stuff to the map.  Could not be any easier.  The display isn't
quite as nice, but guide-key/which-key helps some.

I don't see myself totally removing hydras, just as I find I want to
move a set of keybindings around.