Attempting to make ansi-term "pleasant"
ansi-term attempts to be a very close recreation of a terminal.  In
that spirit, a lot of emacs keys are re-bound or not available.  I
don't seem to be able to do anything with C-w (Vim/Evil window
management prefix) which is frustrating because... I'd like to be able
to leave the window without having to fight with C-x o or use the
mouse.

Well, I'm going to try something else, using M-j...