Hydra command to handle window management
This is intended as a semi-replacement for the C-w commands in Evil,
which are much better than the Emacs defaults (C-x 0/1/2/3) but are
not available in all modes.  The intent here is that all you gotta
do is push M-j s or SPC s.  The 's' is for "split."

It is a hydra.  I said I wasn't a fan of hydras for text menus
consisting of one-shot commands (e.g. you run the command, and it
leaves the menu).  I still feel that way.  But hydras work pretty
nicely for persistent bindings, which is what this does.  You have to
push a non-hydra button to leave, like 'q'.

To split the window vertically, press '|' (or '\').
To split horizontally, press '-'.
These are mnemonic; a horizontal split looks like a '-'; vertical
looks like '|'.  These are the keys I use in tmux as well.

Vim movement keys move from window to window.  Tab does other-window;
's' lets you use ace-window.  Lowercase 'c' closes the current window,
'C' all windows in the frame other than the current window.