Completely removed session management
Or what was there of it... desktop-mode proved to be annoying a lot of
the time.  I really just needed emacs to drop its state somewhere, and
then try to reload it.  But if it needed human intervention in the
process (as it does enough times to be annoying) I would prefer to
just not deal with it, or use something that just saves state to some
files and let you manually load it later.

Furthermore, it also didn't exclude e.g. tramp buffers, which caused a
few problems, notably making the whole loading process slow.  I'm sure
I could have customized it to exclude tramp, but at this point it's
just a pain.

Finally, I wish there was some kind of midnight-mode where it'd knock
out buffers if they hadn't been opened in some time (maybe default to
a fortnight or so?) because I got to the point where it was
autoloading 200 buffers and starting emacs would literally take a few
minutes in the morning.

I may try emacs-session at some point, because I do want some way to
save what files I was working on.  Desktop was just too heavyweight
for me.

There's a pretty high chance that I am going to switch from emacs to
PHPStorm to handle the stuff that pays my bills... and stuff like this
is kind of why.  I am not huge on IDEs, but I gotta hand it to
JetBrains, their stuff is among the best in that class and isn't
priced badly either.  I'm not dropping emacs entirely... it's great
for old-school C and bash and the like, as well as "exotic" languages
like Haskell and Lisps, and org-mode is great.  But its PHP support
has always been second-class or worse.