Completely reworked how I do org-agenda
I now use only five files, and these won't change.

Furthermore, I've went down to four states and five tags.  The
idea is to try to implement GTD in org-mode.

All incoming stuff goes to inbox.org.  I will evaluate
incoming stuff every day at 9am, and move it to one of
the other files.

Everything that I'm actively working on is in gtd.org.  If a
project gets stuck, it is moved to someday.org.

Stuff is moved out of someday.org and back into gtd.org either
on a weekly review (I'm planning to do it on Saturdays at 8pm)
or when someone reminds me about it.

When stuff is completed, it is moved to finished.org.

reminders.org is a 'tickler' file used to remind me of upcoming
things to do or deal with.  I also put birthdays, holidays, and
the GTD reviews in this file.