Added dumb-jump to emacs
I have been playing with PHPStorm over the past month, and one of the
things I've been loving is the jump-to-definition and jump-back.

Emacs is always going to be at a slight disadvantage to a dedicated
IDE in this kind of realm.  Dedicated IDEs, especially commercial ones
that are focused entirely around one language like the JetBrains ones,
are going to have an intimate knowledge of their language that Emacs
will probably always lack without a ton of work.

That said, this plugin comes close enough without needing to install
too many external programs.  It tries to find a definition in your
project using git-grep/ag/rg (and honestly I'm usually in a git repo,
or have ag/rg installed).  It works OK enough with the PHP I threw at
it.

If it gets annoying I might try ctags or global, but I'm not there
yet.