Start to clean up terminal modes a bit...
I'm going to start with eshell, since it is probably going to be the
most amenable to the spacemacs-like bindings I have everywhere else.

eshell was having a similar problem to something I had earlier with
shell-mode, where I had to hit enter twice to run 'ls'; like earlier,
it was a problem with company-mode.

I added the evil-collection for eshell, but I'm not sure what that
does at the moment.