Added a REST client
Because we don't all love using huge Electron applications to do stuff
like this!

With this, we can make a throwaway request by typing SPC f j and then
giving the file an extension of '.restclient'.  Put in your request,
maybe something like

    GET https://pokeapi.co/api/v2/pokemon/6

and hit C-c C-c and you should now be looking at a JSON dump of
Charizard's statistics.  I just fired up Insomnia to verify that that
was the case, and it took it about 30 seconds to fire up.  I'm a
millennial.  Give me thirty seconds of downtime, and I'll already be
bored and moving on to something else.

(Postman is not much better performance-wise; its interface is worse,
and its licensing is bad too.)