Added golang setup
I'm probably going to play a bit with C++ and with these newfangled
curly-brace languages as well.  Starting with Go, but I've been doing
C++ at work and might try my hand at an application outside of work
too.

C++ is still cantankerous, but it's gotten better with C++11 and then
subsequently as well.  Qt is really nice and has waaaaaayyyyy better
documentation than GTK.

Go feels like curly-brace, compilable Python.  It's not as scripty,
but it stays out of your way about as well.

I haven't played with Rust yet; if it supplants anything, it'd
probably be C++.