Added a "proper" license (ISC)
I'm not going to bother to ask a lawyer what license I should use for
my configuration files, I mean, seriously, ugh.

I'm also not a big fan of licensing them with the GPL.  Doing so feels
very heavy-handed considering the end product here: a bunch of
cobbled-together code that configures a bigger piece of software,
almost all of which was taken from somewhere else online.  As such,
I'm using the most minimal license I can find.  ISC is basically a
distillation of the BSD and MIT licenses.  It also has a warranty
disclaimer, which I like (vs. WTFPL) because I'd rather not get sued
because someone used my emacs config in a nuclear reactor.

The ISC license should be compatible with GPLv3[1].  I intrepret "GPL
compatibility" to mean that GPLed code can co-exist with the more
liberally licensed configuration here[2].  I am not a lawyer,
so... take that for what it is worth.  I looked through other people's
.emacs.d projects[3] and found that at least a few used BSD/MIT while
containing GPL'ed bits, which further makes my mind up.

However, if anyone who is more of a lawyer type wants me to relicense
under the GPL for whatever reason, I'm amenable to that discussion.

[1] https://choosealicense.com/licenses/isc/
[2] https://www.gnu.org/licenses/gpl-faq.html#WhatDoesCompatMean
[3] List here: https://github.com/caisah/emacs.dz