(Attempt to) defer loading a lot of utils
I'm not sure if the image+ stuff and various dired utils are being
deferred properly.  I need to keep reading up on how use-package does
this, and particularly if :after still causes the package to be
loaded.

The emoji stuff isn't being deferred as any document might have an
emoji in it and thus need that package.  But since I strongly prefer
the Unicode ones anyways, it should be sufficient to find an
appropriate font and defer emojify until its functions are needed.