Bookmarking shells; easier bookmark access
Inspiration from [1] which I found googling for 'emacs run on docker'.
It doesn't 100% answer my plan, but it's handy nonetheless.

Accessing bookmarks is SPC f b; this is because it is a generic way of
finding a buffer, rather than anything eshell-specific.  Maybe it
should be in SPC b somewhere but meh.

[1] http://manuel-uberti.github.io/emacs/2017/10/19/docker/