Restructured org-mode configuration
It does mostly the same thing, but most of the config is now loaded
from use-package, and I think it should also be deferred but org-mode
is called early on anyways so I have no real way of knowing.

Removed worf as I think it was designed for people not using evil --
you have to be in insert mode for the keys to do anything -- and it by
default annoyingly overrode the [] keys.