Added evil-string-inflection
string-inflection is a package that lets you change snake_case into
PascalCase or camelCase or lisp-case, so on and so forth.

evil-string-inflection lets you use string-inflection on vi(m) text
objects.  So g~io does string inflection on 'io' (the word object at
point).  It can be repeated with the . command.

Need this because I have a ton of variable names that need to be
transformed from snake_case to camelCase and back...