Try to refresh packages less often
On my work PC, I tend to keep the same emacs daemon up for weeks at a
time if possible.  Restarts are rare, so I don't hit this a lot.

On my laptop, I tend to start emacs much more, and so the performance
hit is worse for doing a package refresh.  Also, the internet
connection is liable to not be as good.

So: I added a function that checks to see if MELPA packages have been
refreshed within the past N days (N = 7 here).