Replace smart-mode-line with doom-modeline
I thought I could get sml to work with minions, but I was wrong about
that!  Time to switch...

Found doom-themes; it's already nicer looking and I haven't installed
all-the-icons yet...

The only problem is that it's pretty tall by default -- because you're
intended to use it with the icons turned on.

I was going to use moody, but it's unavailable on emacs 25.2 (easily
available on Ubuntu 18.04) and I'm not sure I want to be the guinea
pig for getting it working there, so... moving on.