Removed guess-style
It was kinda cantankerous and problematic in a few ways.  Lately it
was throwing up random errors about python indentation when loaded,
not really what I want to see.  But the actual functionality (when it
worked) was decent enough, and will be looking for an alternative.
dtrt-indent may be promising, but still need to actually try it.

(One of the things I was griping about was that it wasn't up on
MELPA.  I need to put my stuff up there so I'm not a gigantic
hypocrite!)