Added a string-inflection hydra
This isn't a great place for this code; I don't have a good place for
programming-specific facilities that don't rely on a specific
language... yet.

I know my hydra overlaps evil-string-inflection.  In theory
g~io...... (so on) does the same thing as [esc][spc]liiiiii
(so on).  The latter will be easier for me to remember; but the latter
is otherwise less obtrusive.