Fix getting refresh boolean backwards
This is very likely a big part of why startup is so slow, or at least
the second fix to the jws/is-refresh-needed call.  If it had been less
than seven days since a refresh, (jws/is-refresh-needed 7) would
return a nil indicating "no need for a refresh" -- which means
jws/package-refreshed-already would be nil, which indicates "I haven't
refreshed already."  Yay boolean goofups!

The other "fix," to the jws/package-refresh-once-a-session defun, is
basically equivalent code, just a little cleaner.