Change some web-mode settings
I probably found some of this annoying back when I first set this up,
but now have used other text editors and have come to appreciate it
more.  In particular I have been wanting auto-closing, and in that
particular way.

(I'm not sure if I'm going to go back on auto-quoting, I give it about
a 50% chance)