Added lsp-mode and PHP configuration
I largely grabbed this config from [1].  I've been wanting to play
with lsp-mode for awhile now.

Right now it is only set up to do PHP.

ac-php was 'ok' but lsp-mode is so much better that it is not even
funny, and that's from about ten minutes of kicking the tires a bit.
I'm using the (unfortunately, not libre and freemium) Intelephense[2]
as the PHP lsp-server.  Even just using the free features of
Intelephense, lsp-mode is much more performant and simply is a better
fit to the rest of emacs.

From what I've seen, it literally raises Emacs from a kludgey
greybeard editor to something much more like VS Code.

[1] https://dev.to/gayanhewa/php-with-lsp-mode-on-emacs-5c77

[2] Site: https://intelephense.com/
    Installation: https://www.npmjs.com/package/intelephense

    While Intelephense is not "free" in the FSF sense, key features
    are free-as-in-beer, and advanced features (not sure if emacs
    lsp-mode supports them) cost $10 (not sure if USD or AUD) for a
    perpetual multi-machine one-person license, which is not bad at
    all.