Kick up the LSP file watch threshold
The default of 1000 is too low for my needs; it always warns me, and
then loads quickly anyways on projects of ~2500 files.

I'm not sure this is the ideal configuration for all languages.  At
some point I'm probably just going to turn on dir-locals, especially
to support multi-servers, and I can move this setting there as well
(I think).

But until then, this setting will do.

I also managed to fix my other issue by putting the logic in a hook
defun run with lsp-mode-hook.  And now my files light up with red and
green squigglies in the appropriate places everywhere telling me the
stuff that needs to be cleaned up.  It's a happy Festivus for the
restuvus!