I think I've fixed my company-mode woes
I've been struggling with company-mode completing on RET.  This is OK
in coding major-modes, a little annoying in shell/repl modes, and
frustrating in prose (org-mode, Markdown, Magit commit messages).
This particular set of changes fall out of irritation that my previous
fix made it so you had to hit enter twice to run 'ls' inside eshell...

These changes fix it so all company-mode functionality is now moved to
the TAB button.  RET does nothing.  It aborts company without
completing and instead it does whatever RET would have normally done
outside of company-mode, exactly as I want it.  This is made possible
by company-tng, which is exactly what I've been looking for but never
found until now.

I haven't tested the yasnippet changes extensively, but I use that
much much less than I use autocomplete.  Not entirely sold on the
keybindings but it's important to not have those conflict with
company-mode as in my experience that's pretty bad juju to deal with.