Limit size of lsp-ui popup
The darned thing refuses to go at-point (or bottom) even though I've
explicitly told it to do so.  Something is overwriting my settings.
Top is frustrating because it frequently covers text that I need to
hack on.  at-point doesn't have this problem.  I don't know why it
does this.  It's frustrating.  If I eval the code after lsp-mode is
run, it's fine.

But for whatever reason I can change the size of the box, so I've made
it a little smaller.  It also makes the box a little less useful
(because now I have to drag out the rodent more often to scroll its
contents).  But would prefer that to it blocking stuff.