The dash package seems "fixed" for now...
There's some kind of bodge that had to be put in to make emacs<27
HTTPS work correctly for a number of packages; it apparently bit
magit (haven't noticed?) but also bit dash.

Zeal used to be a nifty package but I feel like newer versions are
slow and unwieldy.

The dash package isn't perfect (namely, whenever you install a docset,
best to let emacs do its thing for about 10-20 seconds) but it
otherwise does the trick.