Get rid of untabify on backspace
Whyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy

Uh, this appears to be the worst of all worlds (unless you are a
space-only bigot and wish to force your opinion on everyone else for
whatever reason)

Backspace should in general do one of two things:

 - delete one codepoint backwards
 - IFF the system is smart enough, detect if there's some kind of
   indent (whether one \t or tab-width # of spaces) and delete back
   one indent.  But ONLY if and only if the system is smart enough,
   and DON'T TRY TO BE SMART and turn it into spaces or tabs or
   anything if it is truly broken indentation, just delete one
   char in that case...

Suffice it to say, this is a dumb enough choice that I don't even
trust emacs to properly handle indent levels until I go another few
months without having to muck around with my settings.

I've kind of stuck with Emacs because it does have a nice ecosystem.
magit is nice.  So is org-mode.  But these kind of things are a
constant ache in my head that keeps making me look longingly at things
like OniVim and the like and wonder if the grass really isn't greener
on the other side.