Create a combined agenda .ics file
This is to share my org agenda with Google Calendar, more or less
following in a similar vein to [1] which is a worthwhile read on its
own.

I added a call to all relevant agenda files to build this file on save
like so, put the following at the very end of your files:

  * COMMENT Local Variables
  # Local Variables:
  # eval: (add-hook 'after-save-hook 'org-icalendar-combine-agenda-files nil t)
  # End:

Then I put the calendar file up on a web server such that it was
visible to Google Calendar.

(Instead of directly putting your .ics file up on Google, put a
directory named /something.ics.  Inside that directory, put a PHP
script as its index that requires a key, or username/pass, etc.)

[1] https://www.pengmeiyu.com/blog/sync-org-mode-agenda-to-calendar-apps/