Some more nip/tuck adjustments to org-ical export
Holy hell this is a nightmare to configure.

Half the problem is iCal itself, which is a fustercluck of a
format.  (Not that I could do much better myself; this is probably a
nightmarish domain to get right.)

Another half of the problem is Google Calendar itself, which
is... we'll not go into it.