Completely reformatted org-capture templates
The actual content of the templates should not have changed.

I added :empty-lines 1 to most of the capture templates -- this should
help keep me from mangling the file structure when capturing, see [1]
for the problem I'm running into.  If this doesn't work, I'll change
the headings to have a "# end of capture" comment or similar to keep
me from shooting off my own foot

[1] https://emacs.stackexchange.com/questions/53526/org-mode-refiling-gobbles-a-newline-and-absorbs-the-next-heading