Use Pyright for Python completion
I was starting to get very unhappy with the pyls server performance,
and Pyright smokes the doors off of it (in a very limited test of it).
(I resent that they aren't open sourcing Pylance... not cool Microsoft)

Also while M$ is definitely evil, Palantir is about as evil as you'd
imagine a company named for magical seeing eyes used by an evil wizard
and a giant flaming evil helleye to communicate evil plans with each
other, might be.  I'll shack up with the ol' em-dollar-sign any day
over that [censored] Thiel dude and his creepy "secure underneath the
watchful eyes" company.