A few more changes to company
To tell the truth, I was struggling with company again (it was
auto-filling the first hit once it popped up... utterly killing
my productivity) and tried to remove that tng-frontend thing,
I'm not sure what its purpose was.

The setqs are taken from another emacs, figured I'd try them
out and see if it helps company usability.