Fix issue when loading org-roam
As I noted in my previous commit, org-roam loads when I boot (for
some reason, probably due to misunderstanding use-package semantics
that I can't think about at the moment because I really need to be
doing things I actually get paid for...) and so I noticed:

When I start org-roam with org-highlight-latex-and-related enabled,
it barfs part of the way through indexing my zettelkasten with an
error

  org-set-regexps-and-options: Wrong type argument: listp, t

and that completely stops bootup in its tracks for whatever reason.

I'm not sure which note is doing it, but I suspect it is a note I made
on electrical theory that contained some math LaTeX markup.

But I can live without LaTeX being bolded in my text so long as I have
the rest of my emacs working lol