Configure lispyville a little more
So, it is somewhat annoying to get into lispy mode from evil normal
mode -- you have to navigate to the beginning/end of a sexp and go
into insert mode.  The 'additional-movement "key theme" for lispyville
adds evil normal commands that move you to the right place and put you
into insert mode so you can then use lispy keys.

So, for instance, if I was in the middle of the use-package lispyville
sexp in normal mode, and I wanted to insert a new sexp above it, I'd
press M-h which would go into insert mode and put point before the
first parens.  Then I'd press '(' to create a new sexp (and then ']'
to go to its end, and RET to make a newline between our sexp and the
existing use-package call).