Dial the file watcher limit back significantly
I've had lsp-mode performance issues... this is part of it.

To my derp brain, 'file watcher' sounds pretty fundamentally
important, but I'm increasingly thinking that all the file watcher
does is see if there are any changes in any files at any point, and
make it available to the LSP server if so.

Which is to say, if the file *rarely changes* it probably shouldn't be
watched.  Which leads us to the following commits...