LSP's file watchers ignore Composer vendor dir
In a small Laravel project, your own code is the tip of an
iceberg of PHP code where almost all of it (to the tune of 8k files)
is underneath the surface in vendor/.

Most of the time, we aren't changing vendored code, right?

If I need to add/remove/update a library, I'd just reload the server
after running composer.