More vc-* bindings
I found this video on the vc- functions:
https://www.youtube.com/watch?v=SQ3Beqn2CEc

I have to say, these are more powerful than you'd give them credit for
at first.  It can't quite do everything Magit does because Magit is
tied to git, and this has to support everything from c. 1972 SCCS all
the way to hg/git.  But it can do *quite a bit*.  Check out the
"annotate" function on a file that gets a lot of changes!