(Hopefully) Fix issue with pdf-tools daemon
Whenever I pull down a fresh copy of this code and start Emacs,
pdf-tools had been pausing midway through to get me to apt-get poppler
so it can compile its daemon.  This is a Bad Thing, I would like this
to only bother when I actually open up a PDF for the first time.  It's
even worse if I launch emacs as a headless daemon, because now the
apt-get request goes nowhere.

I'm guessing the :magic did *not* imply :defer t, so let's not be so
fancy and just sniff the file extension like every other program ever
does lol