Allow searching org-roam bookmarks inside Emacs
This closes the loop entirely; I can now:

 - make bookmarks inside and outside Emacs
 - search bookmarks inside and outside Emacs

and all is good...

---

Not quite.  I note that the org-roam database strips the protocol from
roam_key entries.  So right now I'm guessing the protocol is https,
both in jws/org-roam-search-bookmarks and in my rofi script.  That's
probably bad.

The 'tired' solution would be to try https, then http, then gopher or
gemini or whatever else.

Another yucky solution that might be effective would be to fetch the
org file itself, then simply re-parse the org file itself and get the
whole URL.

The 'wired' solution would be one of:

 - seeing if org-roam could stop stripping the protocol, presuming
   that there isn't some good reason for doing so, which...
 - see if they're storing the protocol some other way