Move most global bindings to use-package decl
This menu was becoming an annoyance.  I'd add code, and then add
something to this menu.  This seems backwards when I can just define
from the use-package declaration, so the bindings will be next to the
rest of its setup code.

I'm not done, but it's late and I"m getting stupider by the minute,
better to push this to later.