Fix weird projectile-command-map bug
So I've always been able to bind into the projectile-command-map, no
problem.  But I've experienced

 - general complaining about "projectile-command-map" not existing
   or the symbol not existing or something, I don't remember

or

 - out and out binding the projectile-command-map s to the global-map,
   thus causing all insert presses of 's' to launch consult-ripgrep

and I just don't want to fight with it anymore.

Short-term, we're just going to move consult-ripgrep to SPC /, loaded
all the time.  This is a function I use *regularly* so I'm fine with
this.

Long-term... may consider replacing projectile with
find-file-in-project and project.el and the like?